Added files to allow comics to be loaded through a plugin like structure
